requirements.txt
python-multipart
fastapi
uvicorn
pydantic < 2.0.0
tortoise-orm